added JS Docs
