+++ ColorPicker/ColorPicker.module.less
@@ -1,93 +1,71 @@
 // ColorPicker.module.less
 //
-@import '../styles/colors.less';
-@import '../styles/mixins.less';
-@import '../styles/variables.less';
+@import "../styles/mixins.less";
+
+@keyframes shakeColor {
+	0% { transform: translateX(0) }
+	25% { transform: translateX(3px) }
+	50% { transform: translateX(-3px) }
+	75% { transform: translateX(3px) }
+	100% { transform: translateX(0) }
+}
+
+.pickerTabLayout {
+	align-items: center;
+}
 
 .colorPicker {
-	align-items: center;
-	justify-content: space-between;
+	display: flex;
+	height: 660px;
+	justify-content: center;
+	min-width: 800px;
+}
 
-	.colorsRow {
-		width: 522px;
-		padding-top: 9px;
-	}
+.selectedColorColumn {
+	margin-top: 15px;
 
-	.coloredButton {
-		border-radius: 72px;
-		border: 3px solid #ffffff;
-		height: 120px;
-		margin-left: 0;
-		min-width: 120px;
-		padding: 0;
-		transform: none;
-
-		.focus({
-			transform: scale(1.1);
-		});
-
-		.disabled({
-			opacity: @sand-disabled-opacity;
-		})
-	}
-
-	.coloredDiv,
-	.coloredInput {
+	.selectedColor {
 		border-radius: 24px;
-	}
+		border: 9px solid;
+		height: 240px;
+		margin-inline: 0;
+		transition: transform .2s;
 
-	.coloredDiv {
-		display: block;
-		height: 81px;
-		border: 3px solid #ffffff;
-
-		&.button, &.focusExpand {
-			margin: 24px;
+		.selectedColorIcon {
+			transition: transform .2s;
 		}
 
 		.focus({
-			transform: scale(1.1);
+			box-shadow: 0 0 12px 6px inset;
+
+			.selectedColorIcon {
+				transform: scale(1.2);
+			}
 		});
-	}
-
-	.coloredInput {
-		opacity: 0;
-
-		&:hover {
-			cursor: pointer;
-		}
 	}
 }
 
-.colorPopup {
-	height: max-content;
+.favoriteColorsRow {
+	text-align: center;
 
-	.colorPopupHeader {
-		font-size: 66px;
-		text-transform: capitalize;
-		margin: 0 0 0 15px;
+	.favoriteColor {
+		border-radius: 50%;
+		border: 6px solid;
+		transition: transform .2s;
+
+		.deleteButton {
+			position: absolute;
+			z-index: -2;
+		}
+
+		.focus({
+			box-shadow: 0 0 12px 3px inset;
+			transform: scale(1.1);
+		});
 	}
 
-	.closeButton {
-		&.button, &.focusExpand {
-			margin: 0;
-		}
-
-		.icon {
-			font-size: 48px;
-		}
-	}
-
-	.colorPickerSliders {
-		padding: 45px 0;
-
-		.colorSlider {
-			margin: 0 27px;
-		}
-
-		.colorSliderText {
-			font-size: 54px;
-			margin: 0 24px;
-		}
+	.shakeFavoriteColor {
+		animation: shakeColor 0.2s;
+		animation-iteration-count: infinite;
 	}
 }
