+++ ColorPicker/ColorPickerSlider.module.less
@@ -1,3 +1,5 @@
+// ColorPickerSlider.module.less
+//
 @import "../styles/mixins.less";
 
 .sliderPickerContainer {
@@ -24,8 +26,7 @@
 		--sand-progress-bg-color-rgb: transparent;
 		border-radius: 99px;
 		height: 33px;
-		margin-left: 18px;
-		margin-right: 18px;
+		margin: 0 18px;
 
 		.focus({
 			.knob {
