--- ColorPickerPOC/ColorPickerPOC.module.less
@@ -1,69 +1,0 @@
-@import "../styles/mixins.less";
-
-.pickerTabLayout {
-	align-items: center;
-}
-
-.colorPicker {
-	display: flex;
-	height: 660px;
-	justify-content: center;
-	min-width: 800px;
-}
-
-.selectedColorColumn {
-	margin-top: 15px;
-
-	.selectedColor {
-		border-radius: 24px;
-		border: 9px solid;
-		height: 240px;
-		margin-inline: 0;
-		transition: transform .2s;
-
-		.selectedColorIcon {
-			transition: transform .2s;
-		}
-
-		.focus({
-			box-shadow: 0 0 12px 6px inset;
-
-			.selectedColorIcon {
-				transform: scale(1.2);
-			}
-		});
-	}
-}
-
-.favoriteColorsRow {
-	text-align: center;
-
-	.favoriteColor {
-		border-radius: 50%;
-		border: 6px solid;
-		transition: transform .2s;
-
-		.deleteButton {
-			position: absolute;
-			z-index: -2;
-		}
-
-		.focus({
-			box-shadow: 0 0 12px 3px inset;
-			transform: scale(1.1);
-		});
-	}
-
-	.shakeFavoriteColor {
-		animation: shakeColor 0.2s;
-		animation-iteration-count: infinite;
-	}
-}
-
-@keyframes shakeColor {
-	0% { transform: translateX(0) }
-	25% { transform: translateX(3px) }
-	50% { transform: translateX(-3px) }
-	75% { transform: translateX(3px) }
-	100% { transform: translateX(0) }
-}
